correção de erro

+++ src/main/java/com/acme/acmeveiculos/domain/service/CadastroTipoVeiculoService.java
@@ -5,12 +5,16 @@
  */
 package com.acme.acmeveiculos.domain.service;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.acme.acmeveiculos.domain.exception.DomainException;
 import com.acme.acmeveiculos.domain.model.TipoVeiculo;
+import com.acme.acmeveiculos.domain.model.Veiculo;
 import com.acme.acmeveiculos.domain.repository.TipoVeiculoRepository;
+import com.acme.acmeveiculos.domain.repository.VeiculoRepository;
 
 @Service
 public class CadastroTipoVeiculoService {
@@ -18,6 +22,9 @@ public class CadastroTipoVeiculoService {
 	@Autowired
 	private TipoVeiculoRepository tipoVeiculoRepository;
 
+	@Autowired
+	private VeiculoRepository veiculoRepository;
+	
 	public TipoVeiculo salvar(TipoVeiculo tipoVeiculo) {
 		TipoVeiculo tipoVeiculoExistente = tipoVeiculoRepository.findByTipo(tipoVeiculo.getTipo());
 		
@@ -30,6 +37,13 @@ public class CadastroTipoVeiculoService {
 	}
 	
 	public void excluir(Long id) {
+		List<Veiculo> listaVeiculos = veiculoRepository.findByTipoVeiculo(tipoVeiculoRepository.findById(id).get());
+		
+		//valida se existe algum veículo cadastrado com o tipo de veículo que vai excluir
+		if (!listaVeiculos.isEmpty()) {
+			throw new DomainException("Para excluir este tipo de veículo, é preciso excluir todos os veículos deste tipo antes.");
+		}
+		
 		tipoVeiculoRepository.deleteById(id);
 	}
